+++ components/SectionAccordion/SectionAccordion.tsx
@@ -5,7 +5,7 @@ import Accordion from "@mui/material/Accordion"
 import AccordionDetails from "@mui/material/AccordionDetails"
 import AccordionSummary from "@mui/material/AccordionSummary"
 import Typography from "@mui/material/Typography"
-import { useState, useCallback } from "react"
+import { useCallback, useState } from "react"
 
 interface SectionAccordionProps {
   title: string